requirements.txt
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
streamlit-extras
urllib3==1.26.6
altair==4.0
openai==0.27.0
tiktoken
langchain==0.0.154
whisper
